admin product update

+++ lib/ui/landing/components/products/widgets/product_details_widget.dart
@@ -54,7 +54,7 @@ class ProductDetailWidget extends StatelessWidget {
                   ProductDetailsModel(
                     price: double.parse(price),
                     discountedPrice: double.parse(dPrice),
-                    qty: int.parse(qty),
+                    qty: double.parse(qty),
                     type: _selectedType,
                   ),
                 );
